add đủ thứ

+++ phucvu/update.php
@@ -9,43 +9,37 @@
 	                VALUES({$_REQUEST['idMon']},'{$_REQUEST['tenBan']}',{$_REQUEST['slMon']})";
 	        mysqli_query($conn,$sql);
 	        $sql = "UPDATE dsBan
-	                SET Mau = 'red', LamMoi = 3
+	                SET Mau = 'red'
 	                WHERE TenBan IN (SELECT TenBan FROM DatMon
 	                                    WHERE IdMon = {$_REQUEST['idMon']} AND TraBan = 0)";
+	        mysqli_query($conn,$sql);		
+			$sql = "UPDATE dsBan
+	                SET LamMoi = 1
+	                WHERE TenBan = '{$_REQUEST['tenBan']}'";
 	        mysqli_query($conn,$sql);
 		}
 	}
 
 	if(isset($_REQUEST['resetLamMoi'])){
-		if($_REQUEST['LamMoi'] == 3){
-			if($_REQUEST['trang']=='phucvu'){
-				$lammoi = 1;
-			} else $lammoi = 2;
-		} else $lammoi = 0;
-		$sql = "UPDATE dsBan SET LamMoi = {$lammoi} WHERE TenBan = {$_REQUEST['tenban']}";
+		$sql = "UPDATE dsBan SET LamMoi = 0 WHERE TenBan = {$_REQUEST['tenban']}";
 		mysqli_query($conn,$sql);
 	}
 
 	if(isset($_REQUEST['huydatmon'])){
-		$sql = "SELECT * FROM DatMon 
-                WHERE TenBan = '{$_REQUEST['tenBan']}' AND TraBan = 0";
-        $result = mysqli_query($conn,$sql);
-
-        if(mysqli_num_rows($result)==1){
-	        $sql = "UPDATE dsBan
-	                SET Mau = '#f6f6f6'
-	                WHERE TenBan = '{$_REQUEST['tenBan']}'";
+		if(isset($_REQUEST['bepdanhan'])){
+			$sql = "UPDATE DatMon SET TrangThai = 5
+					WHERE IdDatMon='{$_REQUEST['idDMon']}'";
+			mysqli_query($conn,$sql);
+		} else {
+	        //delete the record
+	        $sql = "DELETE FROM DatMon
+	                WHERE IdDatMon='{$_REQUEST['idDMon']}'";
         	mysqli_query($conn,$sql);
-        }
+		}
 
         $sql = "UPDATE dsBan
         		SET LamMoi = 1
         		WHERE TenBan IN (SELECT TenBan FROM DatMon WHERE IdMon = {$_REQUEST['idMon']})";
-        mysqli_query($conn,$sql);
-
-        //delete the record
-        $sql = "DELETE FROM DatMon
-                WHERE IdDatMon='{$_REQUEST['idDMon']}'";
         mysqli_query($conn,$sql);
 	}
